requirements.txt
# openseespy-edu-peru — pinned runtime dependencies
#
# OpenSeesPy is the core numerical engine. Version 3.6.0 is the last
# release with stable Python 3.12 wheels on Windows; Python 3.13 is
# NOT supported upstream as of 2026-04.

openseespy==3.6.0

# Core scientific stack.
numpy>=1.26,<2.0
scipy>=1.11,<2.0
matplotlib>=3.8,<4.0

# Config + data.
pyyaml>=6.0
pandas>=2.1,<3.0

# Statistics (used by preflight_statistics.py in the extended framework).
statsmodels>=0.14
